Split slow/duplicated tests from run_all.py into run_more.py
Remove 2 tests (duplicated elsewhere with no unique value):
- tee-rank0-solves 1-iter test (just a print option)
- hydro_ef.py (identical to generic_tester's hydro EF)

Move 8 tests to new run_more.py (covered by dedicated unit tests):
- time_one FarmerLinProx and AircondAMA (timing infrastructure)
- APH farmer x2 (covered by test_aph.py)
- farmer seqsampling BM+BPL (covered by test_conf_int_farmer.py)
- hydro without stage2EF (covered by generic_tester)
- aircond seqsampling BM+BPL (covered by test_conf_int_aircond.py)

Also remove time_one function and pandas/datetime imports from
run_all.py since they're no longer used there.
